Edits to accepted changes
Thanks for these suggestions :) 

I'd prefer to keep the readme impersonal - hence not adopting "Our" - and like "allowing users" more than "to allow the users," so I switched that back. I expect you were making a change based on "allowing" being used twice (thanks for pointing that out!), so I changed the second clause. You're right on "re-run", and that actually should not be hyphenated on reference.

All the best!